Create new layer get_relationship
